outlook/requirements.txt
requesocks==0.10.8
requests==2.19.1
selenium==2.53.6
Flask==1.0.2
Flask-API==1.0
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
PyYAML==3.13